source/lambda/authorizer/requirements.txt
PyJWT==2.8.0
requests==2.32.1